Update and rename organize.js to actions/organize.js
